making deprecate popupManager - in progress
